Revert "refactor: React Query로 변경"
