+++ package.json
@@ -17,14 +17,12 @@
     "@radix-ui/react-slot": "^1.2.3",
     "@radix-ui/react-switch": "^1.2.5",
     "@tailwindcss/vite": "^4.1.10",
-    "@tanstack/react-query": "^5.81.5",
     "axios": "^1.10.0",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "lucide-react": "^0.516.0",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
-    "react-error-boundary": "^6.0.0",
     "react-hook-form": "^7.58.1",
     "react-router-dom": "^7.6.2",
     "sonner": "^2.0.5",
