+++ src/App.tsx
@@ -1,9 +1,6 @@
 import { useEffect } from 'react';
 import AppRoutes from './AppRoutes';
 import { useAuthStore } from '@/stores/auth';
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
-
-const queryClient = new QueryClient();
 
 export default function App() {
   const setToken = useAuthStore((state) => state.setToken);
@@ -13,9 +10,6 @@ export default function App() {
     setToken(token);
   }, [setToken]);
 
-  return (
-    <QueryClientProvider client={queryClient}>
-      <AppRoutes />
-    </QueryClientProvider>
-  );
+  return <AppRoutes />;
 }
+
